booking data

+++ src/main/java/org/mazz/rombooservice/repository/BookingLineRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookingLineRepository extends JpaRepository<BookingLine, Long> {
+public interface BookingLineRepository extends JpaRepository<BookingLine, Integer> {
 
 }
